services/match-service/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.25.2
openai==0.27.8
pytest==7.2.0
requests==2.31.0
